Validacion de sesiones en CRUD videojuegos y en obtencion de comentarios

+++ api/web/rutas_videojuegos.py
@@ -1,11 +1,18 @@
 from flask import request, Blueprint, jsonify
 import controlador_videojuegos
+from controlador_token import token_valido_no_baneado
+from funciones_token import token_existente
 from funciones_auxiliares import Encoder
 
 bp = Blueprint('videojuego', __name__)
 
 @bp.route("/",methods=["GET"])
 def videojuegos():
+    token_valido= token_existente(request.json)
+
+    if token_valido == False:
+        return jsonify({"status":"Token no valido"}), 401
+
     respuesta,code= controlador_videojuegos.obtener_videojuegos()
     return jsonify(respuesta), code
 
@@ -13,6 +20,11 @@ def videojuegos():
 
 @bp.route("/<id>",methods=["GET"])
 def videojuego_por_id(id):
+    token_valido= token_existente(request.json)
+
+    if token_valido == False:
+        return jsonify({"status":"Token no valido"}), 401
+
     respuesta,code = controlador_videojuegos.obtener_videojuego_por_id(id)
     return jsonify(respuesta), code
 
@@ -20,6 +32,12 @@ def videojuego_por_id(id):
 
 @bp.route("/",methods=["POST"])
 def guardar_videojuego():
+    
+    token_valido= token_existente(request.json)
+
+    if token_valido == False or token_valido["perfil"] != "admin":
+        return jsonify({"status":"Token no valido"}), 401
+    
     content_type = request.headers.get('Content-Type')
     if (content_type == 'application/json'):
         videojuego_json = request.json
@@ -40,6 +58,12 @@ def guardar_videojuego():
 
 @bp.route("/<int:id>", methods=["DELETE"])
 def eliminar_videojuego(id):
+
+    token_valido= token_existente(request.json)
+
+    if token_valido == False or token_valido["perfil"] != "admin":
+        return jsonify({"status":"Token no valido"}), 401
+
     respuesta,code=controlador_videojuegos.eliminar_videojuego(id)
     return jsonify(respuesta), code
 
@@ -47,6 +71,12 @@ def eliminar_videojuego(id):
 
 @bp.route("/", methods=["PUT"])
 def actualizar_videojuego():
+
+    token_valido= token_existente(request.json)
+
+    if token_valido == False or token_valido["perfil"] != "admin":
+        return jsonify({"status":"Token no valido"}), 401
+
     content_type = request.headers.get('Content-Type')
     if (content_type == 'application/json'):
         videojuego_json = request.json
